+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "lib/PlayStation2-Game-SDKs"]
-	path = lib/PlayStation2-Game-SDKs
+[submodule "GamePacks/PlayStation2-Game-SDKs"]
+	path = GamePacks/PlayStation2-Game-SDKs
 	url = https://github.com/NightFyre/PlayStation2-Game-SDKs
