+++ examples/Rendering/dllmain.cpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "pch.h"
+#include "CDK.h"
 
 static bool g_running;
