+++ examples/Rendering/example_rendering.vcxproj
@@ -145,8 +145,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="..\..\SDK\core_package.cpp" />
-    <ClCompile Include="..\..\SDK\pcsx2_package.cpp" />
+    <ClCompile Include="..\..\CDK.cpp" />
     <ClCompile Include="dllmain.cpp" />
   </ItemGroup>
   <ItemGroup>
