--- framework.h
@@ -1,16 +1,0 @@
-#pragma once
-
-#define WIN32_LEAN_AND_MEAN             // Exclude rarely-used stuff from Windows headers
-// Windows Header Files
-#include <windows.h>
-#include <iostream>
-#include <string>
-#include <vector>
-#include <memory>
-#include <atomic>
-#include <chrono>
-
-
-//	GAMEPAD
-#include <XInput.h>
-#pragma comment(lib, "XInput.lib")
